add a method that allows changing the server selection
there is now an internal method that actually stores the server selection and an event connected to it and the ui to make sure that everything is always in sync.

as we cannot be completely sure that the UI has been opened before, the class connects its own event, so if the UI is not updated, it will still be triggered internally.